(╯°□°)╯︵ ┻━┻ on deserializing to custom structs
flickr-to-go mostly just saves the JSON from each API call to disk
without looking at it (much). Deserializing to custom structs instead
of map[string]interface{} means that we have explicitly specify the
format of everything that we're just passing through. And because
golang's JSON serializer allows missing and extra fields, we'll
silently ignore any fields in the actual response that don't match our
custom structs.

Switching to map[string]interface{} makes the code that has to actually
look at fields more verbose, but it saves a ton of type boilerplate
and relieves us from having to exhaustively test all the fields that
we're just moving from point A to point B.